+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgload (development version)
 
+* `load_all()` now issues a warning if exported objects conflict with objects defined in the global environment (#112)
+
 # pkgload 1.0.2
 
 * `shim_question()` now works for topics from the R base package that are passed with the double colon operator (e.g. `base::min`) (@mdequeljoe, #99).